lesson-17/05/requirements.txt
influxdb3-python
pandas
python-dotenv